When testing (against) tdbc in the build dir, tdbc.tcl is presumed found in $srcdir/library, not in the same dir as the built shared object.
When testing against an installed tdbc,
tdbc.tcl is presumed found in the same dir as the shared object.

This makes the pkgIndex.tcl in the build dir unusable for testing.
Makefile targets affected: test, shell, gdb, valgrind, etc.

This can be a problem for other extensions as well.

One solution has been to have a [package ifneeded script] generated and used by the Makefile.
This is a bit messy and is needed for all affected targets.
Extensions building against tdbc would also have to act accordingly,
which can be argued is not their responsibility.

This solution uses a pkgIndex.tcl that looks for tdbc.tcl in $srcdir/library.
The $srcdir/library path is stripped-out when pkgIndex.tcl is installed.
With this in place, [package require] simply works.
It is, admittedly, a bit ugly but I think it's the better than:

- Use two pkgIndex.tcl files: one to use when testing and one to install.
- The original method: copy file(s) from the library dir into the build dir
  and remember to clean them up.

The implementation on this branch also removes the extra copying and cleaning-up of tdbc.tcl,
which is no longer necessary.
This has been done in this branch for tdbc.

The tdbc-* extensions building against tdbc would simply have to add $(TDBC_BIN_DIR)
to the existing $(TCLLIBPATH) in their Makefiles to work with this new scheme.

This has been done in this branch for tdbc-*.

The remaining tdbc-* Makefile targets (shell, gdb, valgrind, etc.) remain broken
but can be fixed as was done with tdbc.